Basic Movement Update
A little choppy, but workable:

Added Movement
Added a Jump with:
- Coyote Time
- Jump Buffering
- Modifiable Variables

Raycast is used for the grounded check, only looking for the "Ground" layer. This can be modified to everything with another layer which can set us up for walljumps later, though I think that might arise with issues so I didn't set that up here.

Raycasts are used over a trigger zone because depending on the way we do our logic later for walljumps, players might be "grounded" on the wall (which if we use a "slide" mechanic later will make it harder to display that animation).